Create Deltable.js Component
