+++ components/Delproduct.js
@@ -1,7 +1,7 @@
 //ฟังก์ชั่นลบสินค้า //ทำงานกับหน้าproduct.jsx
 import React, { useState } from 'react';
 import styles from "../styles/product.module.css";
-import { deleteproducts } from "../api"; // import ฟังก์ชัน deleteproducts
+import { deleteproducts } from "../api";
 
 const Delproduct = ({ productId, onDelete }) => {
   const [isDeleting, setIsDeleting] = useState(false);
